requirements.txt
django-jet
six
django-rest-framework